nyancat.c: Fix up sprinkles.
In the top left and top right, a pair of sprinkles were occasionally missing;
they are now always present.

In the bottom left, a corner of the poptart was occasionally frosted instead
of cookie; it is now always cookie to match the other corners.

In the bottom left, a sprinkle was occilating between two locations; it was
moved to the more correct location on the right.